feat: add ability to specify endtime to use for night
This will be useful when header datetime is to be ignored altogether
when calculating things like angle of elevation, moon distance, etc.
This might be important if your fit headers don't contain datetimes
as is the case with our newest camera (it does capture datetime but it's
aways the same bogus datetime value).